Extract AnalyzeButtonBar from Game.tsx
